feat[2.0.0-SNAPSHOT]: [1.2]新增实现MyBatis通用创建人填充

+++ pxc-system/pxc-system-api/src/main/java/io/github/panxiaochao/system/application/api/request/fileaccessory/FileAccessoryCreateRequest.java
@@ -73,7 +73,7 @@ public class FileAccessoryCreateRequest {
 	 * 创建人
 	 */
 	@Schema(description = "创建人")
-	private Long createId;
+	private Integer createId;
 
 	/**
 	 * 创建时间
