+++ pxc-system/pxc-system-api/src/main/java/io/github/panxiaochao/system/application/api/request/sysjob/SysJobUpdateRequest.java
@@ -91,7 +91,7 @@ public class SysJobUpdateRequest {
 	 * 创建人
 	 */
 	@Schema(description = "创建人")
-	private Long createId;
+	private Integer createId;
 
 	/**
 	 * 创建时间
