+++ pxc-system/pxc-system-api/src/main/java/io/github/panxiaochao/system/domain/entity/FileAccessory.java
@@ -67,7 +67,7 @@ public class FileAccessory {
 	/**
 	 * 创建人
 	 */
-	private Long createId;
+	private Integer createId;
 
 	/**
 	 * 创建时间
