+++ pxc-system/pxc-system-api/src/main/java/io/github/panxiaochao/system/domain/entity/SysJob.java
@@ -77,7 +77,7 @@ public class SysJob {
 	/**
 	 * 创建人
 	 */
-	private Long createId;
+	private Integer createId;
 
 	/**
 	 * 创建时间
